rename: object_detector to `FoodMenuDetect` screen widget
